Issue number:  813 Submitted by:  Thierry Lach
Addresses portions of issue 813, specifically:

Changing the name of an attribute on a class.
Changing the name of an operation on a class.
Changing the name of a diagram.

Also, while this does not directly correct "Creating a new stereotype in a package",  entering the name on the new stereotype _will_ cause the model to be marked as changed.

This does _NOT_ close issue 813.