fixing issue 5479
these new critic are now being added as default if there is no
configuration for default profiles

changed ProfileManagerImpl so that it disables the critic in the 
persistend configuration when a critic is disabled when the profile that exports it 
is disabled. Without this, sometimes some critics may be enabled even if 
their profiles is not.